remove old agentctl, rename agentctlv2
