+++ cmd/agentctl/cmd/log.go
@@ -5,9 +5,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/ligato/vpp-agent/cmd/agentctl2/utils"
+	"github.com/ligato/vpp-agent/cmd/agentctl/utils"
 
-	"github.com/ligato/vpp-agent/cmd/agentctl2/restapi"
+	"github.com/ligato/vpp-agent/cmd/agentctl/restapi"
 	"github.com/spf13/cobra"
 )
