+++ cmd/agentctl/cmd/put.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/ligato/cn-infra/servicelabel"
 
-	"github.com/ligato/vpp-agent/cmd/agentctl2/utils"
+	"github.com/ligato/vpp-agent/cmd/agentctl/utils"
 	"github.com/spf13/cobra"
 )
