+++ cmd/agentctl/cmd_generator/generate_config.go
