--- cmd/agentctl/impl/doc.go
@@ -1,2 +1,0 @@
-// Package impl implements handlers for agentctl commands.
-package impl
