--- cmd/agentctl/testing/doc.go
@@ -1,2 +1,0 @@
-// Package testing contains agentctl unit tests.
-package testing
