+++ cmd/agentctl/utils/db_write_utils.go
@@ -15,7 +15,6 @@ import (
 	l2 "github.com/ligato/vpp-agent/api/models/vpp/l2"
 	l3 "github.com/ligato/vpp-agent/api/models/vpp/l3"
 	nat "github.com/ligato/vpp-agent/api/models/vpp/nat"
-	"github.com/ligato/vpp-agent/cmd/agentctl/utils"
 
 	linterface "github.com/ligato/vpp-agent/api/models/linux/interfaces"
 	ll3 "github.com/ligato/vpp-agent/api/models/linux/l3"
@@ -63,7 +62,7 @@ func WriteData(db keyval.ProtoTxn, key string, json string) {
 	case strings.Contains(key, ll3.ModelRoute.KeyPrefix()):
 		writelRouteConfigToDb(db, key, json)
 	default:
-		utils.ExitWithError(utils.ExitInvalidInput,
+		ExitWithError(ExitInvalidInput,
 			errors.New("Unknown input key"))
 	}
 }
@@ -75,14 +74,14 @@ func writeACLConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = proto.Unmarshal([]byte(json), vacl)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert acl json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, vacl)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -95,14 +94,14 @@ func writeInterfaceConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), intr)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert interface json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, intr)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -115,14 +114,14 @@ func writeBridgeDomainConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), bridge)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert interface json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, bridge)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -135,14 +134,14 @@ func writeFibTableConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), fib)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert fib json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, fib)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -155,14 +154,14 @@ func writeXConnectConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), xconnect)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert xconnect json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, xconnect)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -175,14 +174,14 @@ func writeARPConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), arp)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert arp json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, arp)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -195,14 +194,14 @@ func writeRouteConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), route)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert route json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, route)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -215,14 +214,14 @@ func writeProxyConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), proxy)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert proxy json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, proxy)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -235,14 +234,14 @@ func writeIPScanneConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), ipscanner)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert ipscanner json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, ipscanner)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -255,14 +254,14 @@ func writeNATConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), natg)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert nat json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, natg)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -275,14 +274,14 @@ func writeDNATConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), dnat)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert dnat json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, dnat)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -295,14 +294,14 @@ func writeIPSecPolicyConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), policy)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert policy json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, policy)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write policy configuration to Etcd - "+err.Error()))
 	}
 }
@@ -315,14 +314,14 @@ func writeIPSecAssociateConfigToDb(db keyval.ProtoTxn, key string, json string) 
 	err = json2.Unmarshal([]byte(json), ipa)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert ipsec association json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, ipa)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write ipsec association configuration to Etcd - "+err.Error()))
 	}
 }
@@ -335,14 +334,14 @@ func writeIPRedirectConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), ipa)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert ip redirect json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, ipa)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write ip redirect configuration to Etcd - "+err.Error()))
 	}
 }
@@ -355,14 +354,14 @@ func writeToHostConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), ipa)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert tohost json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, ipa)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write tohost configuration to Etcd - "+err.Error()))
 	}
 }
@@ -375,14 +374,14 @@ func writelInterfaceConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), lint)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert linux interface json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, lint)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write linux interface configuration to Etcd - "+err.Error()))
 	}
 }
@@ -395,14 +394,14 @@ func writelARPConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), arp)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert linux arp json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, arp)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write linux arp configuration to Etcd - "+err.Error()))
 	}
 }
@@ -415,14 +414,14 @@ func writelRouteConfigToDb(db keyval.ProtoTxn, key string, json string) {
 	err = json2.Unmarshal([]byte(json), route)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed convert linux route json format to protobuf - "+err.Error()))
 	}
 
 	err = writeDataToDb(db, key, route)
 
 	if err != nil {
-		utils.ExitWithError(utils.ExitError,
+		ExitWithError(ExitError,
 			errors.New("Failed write linux route configuration to Etcd - "+err.Error()))
 	}
 }
